added functionality for hiding the slider on photos without effects
