+++ js/main.js
@@ -1,5 +1,5 @@
 import { drawsThumbnails } from './draws-thumbnails.js';
 import './open-full-size-photo.js';
-import './upload-file.js';
+import './upload-file-modal.js';
 
 drawsThumbnails();
